Retarget frameworks
Drop netcoreapp2.1 and 3.1 because they're no longer supported by Microsoft.
Add net6.0 and net6.0-windows, which are the oldest supported Core CLR runtimes.

Also add PublicAPI tracking files.
