+++ src/Xunit.StaFact/Sdk.WindowsDesktop/WpfFactDiscoverer.cs
