+++ src/Xunit.StaFact/Sdk.WindowsDesktop/WpfTheoryDiscoverer.cs
