+++ src/Xunit.StaFact/Xunit.StaFact.csproj
@@ -1,6 +1,6 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
+﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFrameworks>net472;netstandard2.0;netcoreapp3.1</TargetFrameworks>
+    <TargetFrameworks>net472;netstandard2.0;net6.0;net6.0-windows</TargetFrameworks>
     <RootNamespace>Xunit</RootNamespace>
 
     <Description>Run your xunit-based tests on an STA thread with the WPF Dispatcher, a WinForms SynchronizationContext, or even a cross-platform generic UI thread emulation with a SynchronizationContext that keeps code running on a "main thread" for that test.
@@ -9,25 +9,17 @@ Simply use [WpfFact], [WinFormsFact], [StaFact] or the cross-platform [UIFact] o
 
 Theory variants of these attributes allow for parameterized testing. Check out the xunit.combinatorial nuget package for pairwise or combinatorial testing with theories.</Description>
     <PackageTags>STA xunit testing unit WPF WinForms testing</PackageTags>
-
-    <UseWPF Condition=" '$(TargetFramework)' == 'netcoreapp3.1' ">true</UseWPF>
-    <UseWindowsForms Condition=" '$(TargetFramework)' == 'netcoreapp3.1' ">true</UseWindowsForms>
-    <MSBuildWarningsAsMessages>$(MSBuildWarningsAsMessages);NETSDK1105</MSBuildWarningsAsMessages>
-    <MSBuildWarningsAsMessages>$(MSBuildWarningsAsMessages);NETSDK1106</MSBuildWarningsAsMessages>
   </PropertyGroup>
-  <ItemGroup>
-    <Compile Remove="Sdk.Desktop\**" Condition=" '$(TargetFramework)' == 'netstandard2.0' " />
+  <ItemGroup Condition="'$(TargetingWindows)'!='true'">
+    <Compile Remove="Sdk.WindowsDesktop\**" />
+    <Compile Remove="WindowsDesktop\**" />
   </ItemGroup>
-  <ItemGroup Condition=" '$(TargetFramework)' == 'net472' ">
-    <Reference Include="System.Windows.Forms" />
-    <Reference Include="WindowsBase" />
+  <ItemGroup>
+    <AdditionalFiles Include="$(TargetFramework)\PublicAPI.Shipped.txt" />
+    <AdditionalFiles Include="$(TargetFramework)\PublicAPI.Unshipped.txt" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="xunit.extensibility.execution" />
-  </ItemGroup>
-  <ItemGroup>
-    <!-- Suppress the FrameworkReference in our nuspec file so that our package can still restore on Linux+Mac machines.
-         See https://github.com/AArnott/Xunit.StaFact/issues/35 for more info. -->
-    <FrameworkReference Update="Microsoft.WindowsDesktop.App" PrivateAssets="all" />
+    <PackageReference Include="Microsoft.CodeAnalysis.PublicApiAnalyzers" PrivateAssets="all" />
   </ItemGroup>
 </Project>
