+++ test/Xunit.StaFact.Tests/Samples.cs
@@ -19,7 +19,7 @@ public class Samples
     {
         int initialThread = Environment.CurrentManagedThreadId;
         Assert.NotNull(SynchronizationContext.Current);
-        var expectedApartment = RuntimeInformation.IsOSPlatform(OSPlatform.Windows) ? ApartmentState.STA : ApartmentState.Unknown;
+        ApartmentState expectedApartment = RuntimeInformation.IsOSPlatform(OSPlatform.Windows) ? ApartmentState.STA : ApartmentState.Unknown;
         Assert.Equal(expectedApartment, Thread.CurrentThread.GetApartmentState());
 
         await Task.Yield();
@@ -33,13 +33,13 @@ public class Samples
     [Fact]
     public async Task Fact_OnMTAThread()
     {
-        var expectedApartment = RuntimeInformation.IsOSPlatform(OSPlatform.Windows) ? ApartmentState.MTA : ApartmentState.Unknown;
+        ApartmentState expectedApartment = RuntimeInformation.IsOSPlatform(OSPlatform.Windows) ? ApartmentState.MTA : ApartmentState.Unknown;
         Assert.Equal(expectedApartment, Thread.CurrentThread.GetApartmentState());
         await Task.Yield();
         Assert.Equal(expectedApartment, Thread.CurrentThread.GetApartmentState());
     }
 
-#if (NETFRAMEWORK || NETCOREAPP3_1) && !NON_WINDOWS
+#if NETFRAMEWORK && !NON_WINDOWS
 
     /// <summary>
     /// Demonstrates that <see cref="WpfFactAttribute"/> invokes tests expecting an STA thread
