+++ test/Xunit.StaFact.Tests/WindowsDesktop/WpfFactTests.cs
