+++ test/Xunit.StaFact.Tests/WindowsDesktop/WpfTheoryTests.cs
