--- test/Xunit.StaFact.Tests/xunit.runner.json
@@ -1,3 +1,0 @@
-﻿{
-  "methodDisplay": "classAndMethod"
-}
